telegram-bots/weather-today/requirements.txt
python-telegram-bot==20.0
requests
pydantic
python-dotenv